Moved stats up  to main core module.
